// Dunyana.Web/src/app/pages/admin/model/user.ts
export class user
{
        id:number;
        firstName: string;
        lastName: string;
        address: string;
        name: string;
        email: string; 
        mobile: string; 
        country: string;
        userType: string; 
        isActive: string;
        createdDate: string; 
        naqelUserType: number;
        loginStatus: string;
        result: string;
        company: string;
    
}